+++ mayan/apps/documents/links/document_links.py
@@ -8,7 +8,6 @@ from mayan.apps.navigation.classes import Link
 from ..icons import (
     icon_document_edit, icon_document_list_recent_access, icon_document_list,
     icon_document_preview, icon_document_properties,
-    icon_document_transformations_clear, icon_document_transformations_clone,
     icon_document_type_change, icon_recent_added_document_list
 )
 from ..permissions import (
@@ -16,20 +15,7 @@ from ..permissions import (
     permission_document_view
 )
 
-link_document_clear_transformations = Link(
-    args='resolved_object.id',
-    icon_class=icon_document_transformations_clear,
-    permissions=(permission_transformation_delete,),
-    text=_('Clear transformations'),
-    view='documents:document_clear_transformations'
-)
-link_document_clone_transformations = Link(
-    args='resolved_object.id',
-    icon_class=icon_document_transformations_clone,
-    permissions=(permission_transformation_edit,),
-    text=_('Clone transformations'),
-    view='documents:document_clone_transformations'
-)
+
 link_document_type_change = Link(
     args='resolved_object.id', icon_class=icon_document_type_change,
     permissions=(permission_document_properties_edit,), text=_('Change type'),
@@ -47,12 +33,6 @@ link_document_list_recent_added = Link(
     icon_class=icon_recent_added_document_list, text=_('Recently added'),
     view='documents:document_list_recent_added'
 )
-link_document_multiple_clear_transformations = Link(
-    icon_class=icon_document_transformations_clear,
-    permissions=(permission_transformation_delete,),
-    text=_('Clear transformations'),
-    view='documents:document_multiple_clear_transformations'
-)
 link_document_multiple_type_change = Link(
     text=_('Change type'), icon_class=icon_document_type_change,
     view='documents:document_multiple_type_change'
@@ -62,11 +42,6 @@ link_document_preview = Link(
     permissions=(permission_document_view,), text=_('Preview'),
     view='documents:document_preview'
 )
-#link_document_print_form = Link(
-#    args='resolved_object.id', icon_class=icon_document_print,
-#    permissions=(permission_document_version_print,), text=_('Print'),
-#    view='documents:document_print_form'
-#)
 link_document_properties = Link(
     args='resolved_object.id', icon_class=icon_document_properties,
     permissions=(permission_document_view,), text=_('Properties'),