+++ mayan/apps/documents/views/document_version_views.py
@@ -1,12 +1,17 @@
 import logging
 
+from django.contrib import messages
 from django.template import RequestContext
 from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _, ungettext
 
+from mayan.apps.converter.layers import layer_saved_transformations
+from mayan.apps.converter.permissions import (
+    permission_transformation_delete, permission_transformation_edit
+)
 from mayan.apps.file_caching.tasks import task_cache_partition_purge
 from mayan.apps.views.generics import (
-    MultipleObjectConfirmActionView, MultipleObjectDeleteView,
+    FormView, MultipleObjectConfirmActionView, MultipleObjectDeleteView,
     SingleObjectCreateView, SingleObjectDetailView,
     SingleObjectEditView, SingleObjectListView
 )
@@ -16,6 +21,7 @@ from ..events import event_document_viewed
 from ..forms.document_version_forms import (
     DocumentVersionForm, DocumentVersionPreviewForm
 )
+from ..forms.misc_forms import PageNumberForm
 from ..icons import icon_document_version_list
 from ..links.document_version_links import link_document_version_create
 from ..models.document_models import Document
@@ -27,7 +33,7 @@ from ..permissions import (
 )
 from ..tasks import task_document_version_export
 
-from .misc_views import DocumentPrintFormView, DocumentPrintView
+from .misc_views import PrintFormView, DocumentPrintView
 
 __all__ = (
     'DocumentVersionCreateView', 'DocumentVersionListView',
@@ -233,7 +239,7 @@ class DocumentVersionPreviewView(SingleObjectDetailView):
         }
 
 
-class DocumentVersionPrintFormView(DocumentPrintFormView):
+class DocumentVersionPrintFormView(PrintFormView):
     external_object_class = DocumentVersion
     external_object_permission = permission_document_version_print
     external_object_pk_url_kwarg = 'document_version_id'
@@ -255,3 +261,113 @@ class DocumentVersionPrintView(DocumentPrintView):
         self.external_object.document.add_as_recent_document_for_user(
             user=self.request.user
         )
+
+
+class DocumentVersionTransformationsClearView(MultipleObjectConfirmActionView):
+    model = DocumentVersion
+    object_permission = permission_transformation_delete
+    pk_url_kwarg = 'document_version_id'
+    success_message = _(
+        'Transformation clear request processed for %(count)d document version.'
+    )
+    success_message_plural = _(
+        'Transformation clear request processed for %(count)d document versions.'
+    )
+
+    def get_extra_context(self):
+        result = {
+            'title': ungettext(
+                singular='Clear all the page transformations for the selected document version?',
+                plural='Clear all the page transformations for the selected document version?',
+                number=self.object_list.count()
+            )
+        }
+
+        if self.object_list.count() == 1:
+            result.update(
+                {
+                    'object': self.object_list.first(),
+                    'title': _(
+                        'Clear all the page transformations for the '
+                        'document version: %s?'
+                    ) % self.object_list.first()
+                }
+            )
+
+        return result
+
+    def object_action(self, form, instance):
+        try:
+            for page in instance.pages.all():
+                layer_saved_transformations.get_transformations_for(
+                    obj=page
+                ).delete()
+        except Exception as exception:
+            messages.error(
+                message=_(
+                    'Error deleting the page transformations for '
+                    'document_version: %(document_version)s; %(error)s.'
+                ) % {
+                    'document_version': instance, 'error': exception
+                }, request=self.request
+            )
+
+
+class DocumentVersionTransformationsCloneView(ExternalObjectMixin, FormView):
+    external_object_class = DocumentVersion
+    external_object_permission = permission_transformation_edit
+    external_object_pk_url_kwarg = 'document_version_id'
+    form_class = PageNumberForm
+
+    def dispatch(self, request, *args, **kwargs):
+        results = super().dispatch(request=request, *args, **kwargs)
+        self.external_object.document.add_as_recent_document_for_user(
+            user=request.user
+        )
+
+        return results
+
+    def form_valid(self, form):
+        try:
+            layer_saved_transformations.copy_transformations(
+                delete_existing=True, source=form.cleaned_data['page'],
+                targets=form.cleaned_data['page'].siblings.exclude(
+                    pk=form.cleaned_data['page'].pk
+                )
+            )
+        except Exception as exception:
+            if settings.DEBUG:
+                raise
+            else:
+                messages.error(
+                    message=_(
+                        'Error cloning the page transformations for '
+                        'document version: %(document_version)s; %(error)s.'
+                    ) % {
+                        'document_version': self.external_object,
+                        'error': exception
+                    }, request=self.request
+                )
+        else:
+            messages.success(
+                message=_('Transformations cloned successfully.'),
+                request=self.request
+            )
+
+        return super().form_valid(form=form)
+
+    def get_form_extra_kwargs(self):
+        return {
+            'instance': self.external_object
+        }
+
+    def get_extra_context(self):
+        context = {
+            'object': self.external_object,
+            'submit_label': _('Submit'),
+            'title': _(
+                'Clone page transformations of document version: %s'
+            ) % self.external_object,
+        }
+
+        return context